chore: use @arcblock/eslint-config
