+++ .eslintrc.js
@@ -1,7 +1,6 @@
 module.exports = {
   parser: 'babel-eslint',
-  plugins: ['import'],
-  extends: 'airbnb',
+  extends: ['@arcblock/eslint-config'],
   env: {
     es6: true,
     browser: true,
@@ -9,47 +8,18 @@ module.exports = {
     mocha: true,
     jest: true,
   },
-  rules: {
-    'react/jsx-filename-extension': 'off',
-    'react/jsx-closing-bracket-location': 'off',
-    'jsx-a11y/no-static-element-interactions': 'off',
-    'jsx-a11y/click-events-have-key-events': 'off',
-    'react/forbid-prop-types': 'off',
-    'react/no-unescaped-entities': 'off',
-    'react/sort-comp': 'off',
-    'object-curly-newline': 'off',
-    'operator-linebreak': 'off',
-    'class-methods-use-this': 'off',
-    'max-len': [
-      'error',
-      {
-        code: 120,
-        ignoreComments: true,
-        ignoreUrls: true,
-        ignoreStrings: true,
-        ignoreTemplateLiterals: true,
-        ignoreRegExpLiterals: true,
-      },
-    ],
-    'arrow-parens': ['error', 'as-needed'],
-    'comma-dangle': [
-      'error',
-      {
-        arrays: 'always-multiline',
-        objects: 'always-multiline',
-        imports: 'always-multiline',
-        exports: 'always-multiline',
-        functions: 'ignore',
-      },
-    ],
-    'no-param-reassign': [
-      'error',
-      {
-        props: false,
-      },
-    ],
-  },
   globals: {
     logger: true,
+  },
+  rules: {
+    'unicorn/filename-case': [
+      'error',
+      {
+        cases: {
+          kebabCase: true,
+        },
+        ignore: ['setupProxy.js', 'App.test.js'],
+      },
+    ],
   },
 };
