+++ api/libs/nft/display/index.js
@@ -12,7 +12,10 @@ const getNftName = ({ type }) => {
   return names[key];
 };
 
-const create = (vc, data) => `<svg width="100%" height="100%" viewBox="0 0 374 130" version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink">
+const create = (
+  vc,
+  data
+) => `<svg width="100%" height="100%" viewBox="0 0 374 130" version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink">
   <rect rx="4" ry="4" width="374" height="130" stroke="none" fill="#2f3036" />
   <svg x="3" y="3" width="368" height="124" fill="none">
     <defs>
