+++ api/routes/auth/acquire-asset.js
