+++ api/routes/auth/claim-create-did.js
