+++ api/routes/auth/claim-multiple-step.js
@@ -3,7 +3,6 @@ const { getRandomMessage } = require('../../libs/util');
 
 const { wallet } = require('../../libs/auth');
 
-
 module.exports = {
   action: 'claim_multiple_step',
   claims: [
