+++ api/routes/auth/claim-multiple.js
