+++ api/routes/auth/claim-profile.js
