+++ api/routes/auth/claim-signature.js
