+++ api/routes/auth/consume-vc.js
