+++ api/routes/auth/exchange-asset.js
