+++ api/routes/auth/issue-badge-asset.js
