+++ api/routes/auth/issue-badge.js
