+++ api/routes/auth/nft-private-action.js
