+++ api/routes/auth/nft-private-status.js
