+++ api/routes/auth/verify-nft.js
