+++ src/App.js
@@ -20,7 +20,7 @@ import getWebWalletUrl from './libs/util';
 
 const GlobalStyle = createGlobalStyle`
   a {
-    color: ${(props) => props.theme.colors.green};
+    color: ${props => props.theme.colors.green};
     text-decoration: none;
   }
   a:hover,
@@ -49,7 +49,7 @@ export const App = () => (
           if (session.user) {
             return (
               <UserProvider>
-                <React.Fragment>
+                <>
                   <CssBaseline />
                   <GlobalStyle />
                   <div className="wrapper">
@@ -60,7 +60,7 @@ export const App = () => (
                       <Redirect to="/" />
                     </Switch>
                   </div>
-                </React.Fragment>
+                </>
               </UserProvider>
             );
           }
