+++ src/components/auth/transfer-asset-in.js
@@ -1,7 +1,7 @@
 /* eslint-disable react/jsx-one-expression-per-line */
 import React from 'react';
+// import PropTypes from 'prop-types';
 import useToggle from 'react-use/lib/useToggle';
-import PropTypes from 'prop-types';
 
 import Auth from '@arcblock/did-react/lib/Auth';
 import Button from '@arcblock/ux/lib/Button';
@@ -9,36 +9,31 @@ import Button from '@arcblock/ux/lib/Button';
 import api from '../../libs/api';
 import getWebWalletUrl from '../../libs/util';
 
-export default function TransferTokenAssetIn({ token }) {
+export default function TransferTokenIn() {
   const [isOpen, setOpen] = useToggle(false);
   const webWalletUrl = getWebWalletUrl();
-
   return (
-    <React.Fragment>
+    <>
       <Button color="secondary" variant="contained" size="large" className="action" onClick={() => setOpen(true)}>
-        Send 1 {token.local.symbol} + 1 Asset to Wallet
+        Send 1 Certificate to Wallet
       </Button>
       {isOpen && (
         <Auth
           responsive
-          action="transfer_token_asset_in"
+          action="transfer_asset_in"
           checkFn={api.get}
           socketUrl={api.socketUrl}
           onClose={() => setOpen()}
           onSuccess={() => window.location.reload()}
           messages={{
             title: 'Transfer Required',
-            scan: 'Scan QR code to complete transfer',
+            scan: 'Scan QR code to complete Certificate transfer',
             confirm: 'Confirm on your DID Wallet',
-            success: 'Transfer sent!',
+            success: 'Certificate transfer sent!',
           }}
           webWalletUrl={webWalletUrl}
         />
       )}
-    </React.Fragment>
+    </>
   );
 }
-
-TransferTokenAssetIn.propTypes = {
-  token: PropTypes.object.isRequired,
-};
