+++ src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 // eslint-disable-next-line import/no-named-as-default
-import App from './App';
+import App from './app';
 
 ReactDOM.render(<App />, document.getElementById('root'));
